==> Fall21 -  Ask a Philosopher/Database/EmTechLab/EmTechLab/Notebooks - EmTech/Step1.ipynb <==
### **Hello!**
In this notebook, we will be using the BeautifulSoup package to gather some information about a talk of Krishnamurti and store it in a dataframe.

The website we will be scraping from is: https://jkrishnamurti.org/jksearch 


**Installing Dependencies**

import requests                     # To make 'get' requests through chrome browser
from bs4 import BeautifulSoup       # To parse html file in python tree object
import pandas as pd                 # To create Dataframe and save data into JSON file

import random
import time
from pprint import pprint

from google.colab import data_table
from vega_datasets import data
data_table.enable_dataframe_formatter()

**Getting the Hyperinks**

In this section, we will gather a list of html links to the text documents in the Krishnamurti Archive.

Plan of attack - iterate over the 10  [query pages](https://jkrishnamurti.org/jksearch?keyword=&type=16616&media_type=16616), collecting all of the hyperlinks.

urls = []

# There are 37 query pages we will want to parse and retrieve hyperlinks from

base_url = "https://jkrishnamurti.org/jksearch?keyword=&page={}&talk_type=41&media_type=16616"
search_pages = [base_url.format(str(i)) for i in range(1,38)]
search_pages

"""
# Getting hyperlink paths on the first page
search_page = search_pages[0]

page = requests.get(search_page)
soup = BeautifulSoup(page.content, 'html.parser')

anchor_tags = soup.select(".title > a")
paths = [anchor_tag['href'] for anchor_tag in anchor_tags]

# Notice how the hyperlink Path is relative to the current html website we are on
pprint(paths)
"""

%%capture
# Getting hyperlinks from all query pages - running this cell should take approximately 3-4 minutes
speech_paths = []

for page in search_pages:
  time.sleep(random.randint(1,10))
  page = requests.get(page)
  soup = BeautifulSoup(page.content, 'html.parser')
  anchor_tags = soup.select(".title > a")
  paths = [anchor_tag['href'] for anchor_tag in anchor_tags]
  speech_paths.append(paths)

pprint(speech_paths)

# Turning paths into hyperlinks
host_name = "https://jkrishnamurti.org/"
urls = []

for page_paths in speech_paths:
  for path in page_paths:
    urls.append(host_name + path)

pprint(urls)

**Webscraping**

In this section, we will inspect a single speech path, and collect the speech and metadata from it, and create a dataframe with single row.  
  
Next, we repeat the process for all urls we collected in previous section

# Making a 'get' request
page = requests.get("https://jkrishnamurti.org/content/action-without-conflict")
soup = BeautifulSoup(page.content, 'html.parser')

# Getting <div> with all text of speech
text = [paragraph.get_text() for paragraph in soup.find('div', property="content:encoded").find_all('p')]

# Creating speech string
speech = ' '.join(text)

# Getting the <div> with all the fields
fields = soup.select(".custom-footer .field-wrapper")

# Get field lables and values
f_label = fields[0].select(".field-label")[0].get_text().strip()
f_value = fields[0].find('div').get_text().strip()

labels = [field.select(".field-label")[0].get_text().strip() for field in fields]
values = [field.find('div').get_text().strip() for field in fields]
values = [[value] for value in values]

# Creating a dictionary containing text and metadata
d = dict(zip(labels, values))
d['text'] = speech

# Creating dataframe with one row
speech_df = pd.DataFrame.from_dict(d)

speech_df

# Here we want to make to have a progress bar
count = 0
# increment count, everytime count % 10 == 0 print out a bar

# Let us add another row to this dataframe
for url in urls:
  # Print bar
  if count % 10 == 0:
    print(#, end="")
    if count % 100 == 0:
      print()
  # Halt program for random period of time
  time.sleep(random.randint(1, 60))

  # Make 'get' request
  page = requests.get(url)
  soup = BeautifulSoup(page.content, 'html.parser')

  text = [paragraph.get_text() for paragraph in soup.find('div', property="content:encoded").find_all('p')]
  speech = ' '.join(text)

  # Getting the <div> with all the 
  fields = soup.select(".custom-footer .field-wrapper")

  # Creating a dataframe with speech and metadata
  labels = [field.select(".field-label")[0].get_text().strip() for field in fields]
  values = [field.find('div').get_text().strip() for field in fields]
  d = dict(zip(labels, values))
  d['text'] = speech
  speech_df = speech_df.append(d, ignore_index = True)

speech_df

**Saving dataframe to JSON**

speech_df.to_json("krishnamurti.json")

# This will download json to your computer
from google.colab import files
files.download("krishnamurti.json")
