==> Fall21 -  Ask a Philosopher/Database/EmTechLab/EmTechLab/Notebooks - EmTech/Step2.ipynb <==
# **Step 2**
In this notebook, we will be using Pandas to clean the Krishnamurti's speeches


import pandas as pd

# Formatting pandas DataFrame display
from google.colab import data_table
data_table.disable_dataframe_formatter()

# Reading in JSON
df = pd.read_json("krishnamurti.json")
df.sort_values('text').head(4)


df

**Text**  
Thing to clean:


*   Numerical Dates prepending each string
*   Speaker identification tags
*   Page tags



# Numerical Dates
df['text'] = df['text'].str.replace(r"(\d)", "")

# Speaker identification tags
df['text'] = df['text'].str.replace(r"(\w)*(:)", "")

# Page tags
df['text'] = df['text'].str.replace(r"-- Page--", "")
df['text'] = df['text'].str.replace(r"- END -", "")
df['text'] = df['text'].str.replace(r"Page Page", "")
df['text'] = df['text'].str.replace(r"Page(of )", "")


# We should see no numbers, page tags, or speaker identification
df.sort_values('text').head(4)

**Great!**  
Let's now load out cleaned speeches into a new json file

df.to_json("krishnamurti_clean.json")
