==> Fall21 -  Ask a Philosopher/Database/EmTechLab/EmTechLab/Notebooks - EmTech/Step3.ipynb <==
# Step 3
Here we will load the JSON file containing Krishnamurti's speeches, and work with it as a PySpark dataframe.

Related resources:

*   [PySpark in Colab](https://medium.com/linkit-intecs/pyspark-with-google-colab-d964fd693ca7)

*   [PySpark Documentation](https://spark.apache.org/docs/3.1.3/api/python/getting_started/quickstart.html)

* [Spark SQL Basics](https://towardsdatascience.com/pyspark-and-sparksql-basics-6cb4bf967e53)



**Installing Dependencies**

from google.colab import drive
drive.mount('/content/drive')

%%capture
# Install pyspark module (with additional SQL functionality)
!pip install pyspark[sql]

from pyspark.sql import SparkSession
spark = SparkSession.builder\
        .master("local")\
        .appName("Colab")\
        .config('spark.ui.port', '4050')\
        .getOrCreate()

import pandas as pd

# Formatting pandas DataFrame display
from google.colab import data_table
data_table.enable_dataframe_formatter()

**Create PySpark DataFrame**  
We will first have to upload `krishnamurti.json`, before running this code block

# Reading in JSON
pandas_df = pd.read_json("krishnamurti_clean.json")

# Creating Spark DataFrame
df = spark.createDataFrame(pandas_df)

# Configuring dataframe (for notebooks only)
spark.conf.set('spark.sql.repl.eagerEval.enabled', True)
spark.conf.set('spark.sql.repl.eagerEval.maxNumRows', 10)

df.printSchema()

# Removing duplicates from dataframe
df = df.drop_duplicates(['text'])

df

df.select("Participants Category", "Country", "Decade").describe().show()

# Creating a groupby object
df.groupby('City').count().show()

# Running SQL queries
df.createOrReplaceTempView("krishnaTable")
spark.sql("SELECT count(*) as number_of_speeches from krishnaTable").show()

# Can register and invoke user-defined functions
from pyspark.sql.functions import pandas_udf
from pyspark.sql.functions import expr

@pandas_udf("string")
def fix_decades(s: pd.Series) -> pd.Series:
    return s

spark.udf.register("fix_decades", fix_decades)

df.selectExpr('fix_decades(Decade)').show()

# Finding the number of texts beginning with digitis
df[df.text.rlike(r"life")].describe().show()

# Looking at attributes and methods of this df column
dir(df.text)

# Changing names and dropping unimportant columns
df = df.withColumnRenamed('text', 'Speech')
df.drop("Text source", "Talk Type", "Participants Category", "Date Code", "Participants")

# Saving clean version
df.select("Decade", "Country", "City", "Speech") \
.write \
.save("spark_krishanmurti.json",format="json") # This should be saved in a file named"part-0000...json"

# Ending spark session
spark.stop()
