==> Fall21 -  Ask a Philosopher/Database/EmTechLab/EmTechLab/Notebooks - EmTech/step5.ipynb <==
In this notebook, I will create a framework for a demo of this language model using Gradio's Interface and Blocks class

**Installing Dependencies**

%%capture
# Install Huggingface Transformers
!pip install transformers
!pip install gradio
from transformers import pipeline

# Loading in Language Model
my_model = "eliwill/distilgpt2-finetuned-final-project"
krishnamurti_generator = pipeline("text-generation",my_model)

# importing 
import gradio as gr

###### **Gradio Interface**
In this section, we will use Gradio's high-level `Interface` class to create the framework of our demo.

The interface class requires three parameters:
>
>1. fn = Function UI built around
2. Inputs = List of input elements
3. Outputs = List of output elements


**fn**

"""This function will return:
(1) An answer to the question with the syntax of Krishnamurti
(2) Dataframe containing related articles and quotes of Krishnamurti"""
def ask_krishnamurti(question):
  answer = krishnamurti_generator(question)[0]['generated_text']
  return answer, df.head(5)

**outputs**

# Loading in DataFrame -- make sure to import in Colab session
import pandas as pd
df = pd.read_json("krishnamurti_clean.json")

df = df.drop(columns=["Text source", "Talk Type", "Participants Category", "Decade", "Participants", "Date Code"])

demo = gr.Interface(fn=ask_krishnamurti, 
                    title="Ask Krishnamurti",
                    description="A Q/A language model that responds to queries using the syntax of Jiddu Krishnamurti",
                    inputs=[gr.Textbox(
                                        placeholder="Ask question here!", 
                                        lines=5, 
                                        max_lines=10,
                                        label="Question")],
                    outputs=[gr.Textbox(
                        lines=3, 
                        max_lines=10,
                        label="Answer"
                    ),  
                        gr.DataFrame(
                        headers=["Country","City", "Speech"],
                        max_rows=2,
                        interactive=True,
                        wrap=False)],
                    examples=[["Why do we feel anger?", True, True]]
                    )

# Launching demo
demo.launch(inbrowser=True)
